DEV-16025 Use declarative loading strategy API
This adds configuration for the new hierarchyChildNodesQuery and
hierarchyContentQuery properties with the aim to maintain maximum
compatibility with DITA and keep the configuration as simple and
footgun-free as possible. To that end, this is fully based on the
fonto:dita-class function. I've added a note on the associated
performance concerns, which should only be necessary in the more extreme
cases.