DEV-19021 Ensure mathml-edit is used with the correct contextNodeId.
mathml-edit should point to a mathml element, not a container with
(1 or more) mathml elements in it, like an equation-block or -inline.
Otherwise it will always edit the first mathml element in them, and not
the element your actually right-clicking on.

So also no longer disable contextual operations on mathml elements that
are in an inline layout context.